Add support to explicitly set the enabled SSL protocols
For some reason when the server uses SSLv3 the lib is not able to decode
the messages properly. To workaround this issue it's needed to set the
enabled protocols to SSLv3 (only).

This patch enables to explicitly set the enabled protocols.

For more info: https://github.com/koush/AndroidAsync/issues/174