fix #181 configurable reconnectDelay and reconnectDelayMax for socket.io
Add class SocketIORequest.Config whose instances may optionally be passed to SocketIORequest's constructor to configure reconnectDelay (initial delay) and reconnectDelayMax.  The delay will double on each attempt as before, but never exceed reconnectDelayMax, if specified.  If reconnectDelayMax is 0 (the default), there is no max and reconnect delay will grow forever (as it did before this change).